+++ patches/GameDirectorPatch.cs
@@ -16,6 +16,16 @@ static class GameDirectorPatch
 
     [HarmonyPrefix, HarmonyPatch(nameof(GameDirector.gameStateStart))]
     private static void GameStateStart_Prefix() {
+        if (SemiFunc.IsMasterClient() && Configuration.hideValuables.Value) {
+            NetworkController controller = Map.Instance.gameObject.GetComponent<NetworkController>();
+            if (controller == null) {
+                DeadMap.Logger.LogWarning("Failed to send hide valuables event: Network controller is null.");
+                return;
+            }
+
+            controller.photonView.RPC("HideValuables", Photon.Pun.RpcTarget.All, new object[0]{});
+        }
+
         if (DeadMap.spectating) {
             DeadMap.SetSpectating(false);
         }